refactor(ir): move list comprehension/folding to SymbolTable
